Migrate extras to core
