+++ youtubesearchpython/__future__/streamurlfetcher.py
@@ -1,8 +1,8 @@
 from typing import Union
-from youtubesearchpython.__future__.internal.streamurlfetcher import StreamURLFetcherInternal
+from youtubesearchpython.core.streamurlfetcher import StreamURLFetcherCore
 
 
-class StreamURLFetcher(StreamURLFetcherInternal):
+class StreamURLFetcher(StreamURLFetcherCore):
     '''Gets direct stream URLs for a YouTube video fetched using `Video.get` or `Video.getFormats`.
 
     This class can fetch direct video URLs without any additional network requests (that's really fast).
@@ -52,8 +52,8 @@ class StreamURLFetcher(StreamURLFetcherInternal):
             >>> print(url)
             "https://r6---sn-gwpa-5bgk.googlevideo.com/videoplayback?expire=1610798125&ei=zX8CYITXEIGKz7sP9MWL0AE&ip=2409%3A4053%3A803%3A2b22%3Adc68%3Adfb9%3Aa676%3A26a3&id=o-APBakKSE2_eMDMegtCmeWXfuhhUfAzJTmOCWj4lkEjAM&itag=251&source=youtube&requiressl=yes&mh=aP&mm=31%2C29&mn=sn-gwpa-5bgk%2Csn-gwpa-qxad&ms=au%2Crdu&mv=m&mvi=6&pl=36&initcwndbps=146250&vprv=1&mime=audio%2Fwebm&ns=ULL4mkMO31KDtEhOjkOrmpkF&gir=yes&clen=10210834&dur=634.601&lmt=1544629945422176&mt=1610776131&fvip=6&keepalive=yes&c=WEB&txp=5511222&n=uEjSqtzBZaJyVn&sparams=expire%2Cei%2Cip%2Cid%2Citag%2Csource%2Crequiressl%2Cvprv%2Cmime%2Cns%2Cgir%2Cclen%2Cdur%2Clmt&sig=AOq0QJ8wRAIgKKIEiwQTgXsdKPEyOckgVPs_LMH6KJoeaYmZic_lelECIHXHs1ZnSP5mgtpffNlIMJM3DhxcvDbA-4udFFE6AmVP&lsparams=mh%2Cmm%2Cmn%2Cms%2Cmv%2Cmvi%2Cpl%2Cinitcwndbps&lsig=AG3C_xAwRQIhAPmhL745RYeL_ffgUJk_xJLC-8riXKMylLTLA_pITYWWAiB2qUIXur8ThW7cLfQ73mIVK61mMZc2ncK6FZWjUHGcUw%3D%3D"
         '''
-        await self._getDecipheredURLs(videoFormats)
-        for stream in self.player_response["url_encoded_fmt_stream_map"]:
+        self._getDecipheredURLs(videoFormats)
+        for stream in self._streams:
             if stream["itag"] == itag:
                 return stream["url"]
         return None
@@ -137,5 +137,5 @@ class StreamURLFetcher(StreamURLFetcherInternal):
                 ]
             }
         '''
-        await self._getDecipheredURLs(videoFormats)
-        return {"streams": self.player_response["url_encoded_fmt_stream_map"]}
+        self._getDecipheredURLs(videoFormats)
+        return {"streams": self._streams}
