+++ youtubesearchpython/core/componenthandler.py
@@ -18,3 +18,16 @@ def getValue(source: dict, path: List[str]) -> Union[str, int, dict, None]:
                 break
     return value
 
+
+def getVideoId(videoLink: str) -> str:
+    if 'youtu.be' in videoLink:
+        if videoLink[-1] == '/':
+            return videoLink.split('/')[-2]
+        return videoLink.split('/')[-1]
+    elif 'youtube.com' in videoLink:
+        if '&' not in videoLink:
+            return videoLink[videoLink.index('v=') + 2:]
+        return videoLink[videoLink.index('v=') + 2: videoLink.index('&')]
+    else:
+        return videoLink
+
